Add custom exceptions

+++ src/Core/HrLeaveManagement.Application/Features/LeaveType/Commands/DeleteLeaveType/DeleteLeaveTypeCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using HrLeaveManagement.Application.Contracts.Persistence;
+using HrLeaveManagement.Application.Exceptions;
 using MediatR;
 
 namespace HrLeaveManagement.Application.Features.LeaveType.Commands.DeleteLeaveType;
@@ -14,6 +15,7 @@ public class DeleteLeaveTypeCommandHandler(ILeaveTypeRepository leaveTypeReposit
         var leaveTypeToDelete = await leaveTypeRepository.GetByIdAsync(request.Id);
 
         // Verify that record exists
+        if (leaveTypeToDelete is null) throw new NotFoundException(nameof(LeaveType), request.Id);
 
         // Remove from database
         await leaveTypeRepository.DeleteAsync(leaveTypeToDelete);
